requirements.txt
flake8
flake8-bugbear
GitPython
mypy
requests
types-requests